Organize the code somewhat
